requirements.txt
torch==1.9.0
pandas==1.3.3
transformers==4.11.3
absl-py==0.13.0